Group containers in a separate folder.
