+++ source/tit/core/containers/boost.hpp
@@ -1,0 +1,36 @@
+/* ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ *\
+ * Part of the Tit Solver project, under the MIT License.
+ * See /LICENSE.md for license information. SPDX-License-Identifier: MIT
+\* ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ */
+
+#pragma once
+
+#include <functional>
+
+#include <boost/container/container_fwd.hpp>
+#include <boost/container/flat_map.hpp>
+#include <boost/container/small_vector.hpp>
+
+#include "tit/core/basic_types.hpp"
+
+namespace tit {
+
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+/// Boost small vector.
+template<class T, size_t N>
+using SmallVector = boost::container::small_vector<T, N>;
+
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+/// Boost flat map.
+template<class Key, class T, class Compare = std::less<Key>>
+using FlatMap = boost::container::flat_map<Key, T, Compare>;
+
+/// Boost small flat map.
+template<class Key, class T, size_t N, class Compare = std::less<Key>>
+using SmallFlatMap = boost::container::small_flat_map<Key, T, N, Compare>;
+
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+} // namespace tit
