+++ source/tit/core/containers/mdvector.test.cpp
@@ -7,7 +7,7 @@
 #include <array>
 #include <ranges>
 
-#include "tit/core/mdvector.hpp"
+#include "tit/core/containers/mdvector.hpp"
 
 #include "tit/testing/test.hpp"
