+++ source/tit/sph/particle_mesh.hpp
@@ -16,7 +16,7 @@
 
 #include "tit/core/basic_types.hpp"
 #include "tit/core/checks.hpp"
-#include "tit/core/multivector.hpp"
+#include "tit/core/containers/multivector.hpp"
 #include "tit/core/par.hpp"
 #include "tit/core/profiler.hpp"
 #include "tit/core/stats.hpp"
